+++ lib/loader.js
@@ -50,12 +50,6 @@ module.exports = function loader(content, map) {
 
       const alreadyImported = {};
       const importJs = result.importItems
-        .map((imp) => {
-          // fixes #781 when importing `url(filename.css )`
-          // eslint-disable-next-line no-param-reassign
-          imp.url = imp.url.trim();
-          return imp;
-        })
         .filter((imp) => {
           if (!imp.mediaQuery) {
             if (alreadyImported[imp.url]) {
@@ -98,7 +92,11 @@ module.exports = function loader(content, map) {
       // helper for ensuring valid CSS strings from requires
       let urlEscapeHelper = '';
 
-      if (options.url !== false && result.urlItems.length > 0) {
+      if (
+        options.url !== false &&
+        result.urlItems &&
+        result.urlItems.length > 0
+      ) {
         urlEscapeHelper = `var escape = require(${loaderUtils.stringifyRequest(
           this,
           require.resolve('./runtime/escape.js')
