+++ lib/plugins/index.js
@@ -1,0 +1,9 @@
+const importParser = require('./postcss-import-parser');
+const parser = require('./postcss-parser');
+const urlParser = require('./postcss-url-parser');
+
+module.exports = {
+  importParser,
+  parser,
+  urlParser,
+};
