+++ lib/processCss.js
@@ -10,7 +10,7 @@ const extractImports = require('postcss-modules-extract-imports');
 const modulesScope = require('postcss-modules-scope');
 const modulesValues = require('postcss-modules-values');
 
-const cssLoaderParser = require('./postcss-css-loader-parser');
+const { importParser, parser, urlParser } = require('./plugins');
 
 const Warning = require('./Warning');
 const CssSyntaxError = require('./CssSyntaxError');
@@ -26,7 +26,7 @@ module.exports = function processCss(content, map, options, callback) {
     import: loaderOptions.import !== false,
   };
 
-  const pipeline = postcss([
+  const plugins = [
     modulesValues,
     localByDefault({
       mode: loaderOptions.modules ? 'local' : 'global',
@@ -63,10 +63,19 @@ module.exports = function processCss(content, map, options, callback) {
         );
       },
     }),
-    cssLoaderParser(parserOptions),
-  ]);
+  ];
 
-  pipeline
+  if (loaderOptions.import !== false) {
+    plugins.push(importParser(parserOptions));
+  }
+
+  if (loaderOptions.url !== false) {
+    plugins.push(urlParser(parserOptions));
+  }
+
+  plugins.push(parser(parserOptions));
+
+  postcss(plugins)
     .process(content, {
       // we need a prefix to avoid path rewriting of PostCSS
       from: `/css-loader!${loaderUtils
