+++ lib/utils.js
@@ -14,7 +14,7 @@ function dashesCamelCase(str) {
 }
 
 function compileExports(result, importItemMatcher, camelCaseKeys) {
-  if (!Object.keys(result.exports).length) {
+  if (!result.exports || !Object.keys(result.exports).length) {
     return '';
   }
